after some attempt to d=5
after try to brute force for d=5 which didn't complete in 20h, I found that it is impossible to brute force d=5 to get vertices and its generators. I believe it is because it takes more time to store generators at each iteration. So now use the result of last year which only contains vertices d=5

Add more function body in the verificationall() function to verify d=5 due to different file struture

Monitored a bug, cannot use a d for all since the main func needs to import other funcs using d in other files